+++ styleguide.config.js
@@ -26,6 +26,10 @@ module.exports = {
       components: () => ([ path.resolve(__dirname, 'src/components/ui/*', 'index.js') ])
     },
     {
+      name: 'Map',
+      components: () => ([ path.resolve(__dirname, 'src/components/map/*', 'index.js') ])
+    },
+    {
       name: 'Form',
       components: () => ([ path.resolve(__dirname, 'src/components/form/*', 'index.js') ])
     }